requirements.txt
discord.py==1.7.3
lightgbm==3.3.1
numpy==1.21.5
pandas==1.3.5
python-dotenv==0.19.2
scikit-learn==1.0.1
PyNaCl==1.5.0